fix(session): 对四个数据表的数据获取权限加入session验证

+++ transit/apis/stations.py
@@ -204,6 +204,13 @@ def dispatcher(request):
     :param request:
     :return:json
     """
+    if 'usertype' not in request.session:
+        return JsonResponse({
+            'ret': 302,
+            'msg': '未登录',
+            'redirect': 'sign.html'
+        }, status=302)
+
     # 将请求参数统一放在request的params属性中，方便后续处理
     # GET请求 参数在url中，通过request对象的GET属性获取
     if request.method == 'GET':
@@ -218,11 +225,11 @@ def dispatcher(request):
     action = request.params['action']
     if action == 'list_station':
         return list_station(request)
-    elif action == 'add_station':
+    elif action == 'add_station' and request.session['usertype'] == 'admin':
         return add_station(request)
-    elif action == 'modify_station':
+    elif action == 'modify_station' and request.session['usertype'] == 'admin':
         return modify_station(request)
-    elif action == 'del_station':
+    elif action == 'del_station' and request.session['usertype'] == 'admin':
         return delete_station(request)
     else:
         return JsonResponse({'ret': 1, 'msg': '不支持该类型http请求'})
